chat-server/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-socketio==5.13.0
aiomysql==0.2.0
PyJWT==2.10.1
pydantic==2.11.7
python-dotenv==1.1.1
langgraph==0.2.74
python-multipart==0.0.20
